+++ CustomRunOptions/CustomRunOptions/BootSequenceCustomRunOption.cs
@@ -210,7 +210,7 @@ internal sealed class BootSequenceCustomRunOption : ICustomRunOptionsApi.ICustom
 					]),
 					api.MakeBackButton()
 				])))
-			).SetValueText(() => config.GetEnforcedBootSequenceDownside() is { } key && UpsideChoices.TryGetValue(key, out var choice)
+			).SetValueText(() => config.GetEnforcedBootSequenceDownside() is { } key && DownsideChoices.TryGetValue(key, out var choice)
 				? (key == "" ? ModEntry.Instance.Localizations.Localize(["options", nameof(BootSequenceCustomRunOption), "none"]) : choice.Title)
 				: $"<c={Colors.textMain.gain(0.5)}>{ModEntry.Instance.Localizations.Localize(["options", nameof(BootSequenceCustomRunOption), "notSelected"])}</c>"
 			).SetHeight(12).SetTitleFont(() => DB.pinch).SetValueTextFont(() => DB.pinch),